REF: introduce comm module
